requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chainer==3.0.0
chardet==3.0.4
click==6.7
dill==0.2.7.1
filelock==2.0.13
Flask==0.12.2
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
google-api-core==0.1.1
google-auth==1.2.1
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-language==1.0.0
google-cloud-storage==1.6.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.7.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mock==2.0.0
nose==1.3.7
numpy==1.13.3
oauth2client==3.0.0
pbr==3.1.1
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
requests==2.18.4
rsa==3.4.2
six==1.11.0
slackbot==0.5.1
slacker==0.9.60
urllib3==1.22
websocket-client==0.44.0
Werkzeug==0.12.2
